+++ src/chapter1/ducks/behaviours/fly/FlyBehavior.java
@@ -8,5 +8,5 @@ public interface FlyBehavior {
     /**
      * Fly behavior
      */
-    public void fly();
+    void fly();
 }
