+++ src/chapter1/ducks/behaviours/fly/FlyWithWings.java
@@ -1,5 +1,8 @@
 package chapter1.ducks.behaviours.fly;
 
+/**
+ * For the Ducks that fly with wings
+ */
 public class FlyWithWings implements FlyBehavior {
 
     @Override
