+++ src/chapter1/ducks/behaviours/quack/MuteQuack.java
@@ -1,9 +1,12 @@
 package chapter1.ducks.behaviours.quack;
 
+/**
+ * Mute Quack
+ */
 public class MuteQuack implements QuackBehavior {
 
     @Override
     public void quack() {
-        System.out.println("Quack");
+        System.out.println("<< Silence >>");
     }
 }
