+++ src/chapter1/ducks/behaviours/quack/Quack.java
@@ -1,0 +1,12 @@
+package chapter1.ducks.behaviours.quack;
+
+/**
+ * The Quack behavior
+ */
+public class Quack implements QuackBehavior {
+
+    @Override
+    public void quack() {
+        System.out.println("Quack");
+    }
+}
