+++ src/chapter1/ducks/behaviours/quack/Squeak.java
@@ -1,9 +1,12 @@
 package chapter1.ducks.behaviours.quack;
 
+/**
+ * Squeak Behavior
+ */
 public class Squeak implements QuackBehavior {
 
     @Override
     public void quack() {
-        System.out.println("<< Silence >>");
+        System.out.println("Sneak");
     }
 }
